requirements.txt
opencv-python>=4.5.0
openai>=1.0.0
Pillow>=9.0.0
python-dateutil>=2.8.2
numpy>=1.20.0
pytesseract>=0.3.8
python-dotenv>=0.19.0
matplotlib>=3.5.0
requests>=2.27.0 